improved structure of parametrization module
