+++ envinorma/parametrization/models/__init__.py
@@ -1,5 +1,4 @@
-from .models import (  # noqa: F401
-    AlternativeSection,
+from .condition import (  # noqa: F401
     AndCondition,
     Condition,
     Conditions,
@@ -12,10 +11,14 @@ from .models import (  # noqa: F401
     MergeCondition,
     MergeConditions,
     MergeType,
-    NonApplicationCondition,
     OrCondition,
+    Range,
+)
+from .parameter import Parameter, ParameterEnum, ParameterType, parameter_value_to_str  # noqa: F401
+from .parametrization import (  # noqa: F401
+    AlternativeSection,
+    NonApplicationCondition,
     ParameterObject,
     ParameterObjectWithCondition,
     Parametrization,
-    Range,
 )
