+++ tests/test_natural_language_warnings.py
@@ -1,14 +1,9 @@
 from datetime import datetime
 
 from envinorma.models.classement import Regime
-from envinorma.parametrization.conditions import (
-    AndCondition,
-    Equal,
-    Greater,
-    Littler,
-    OrCondition,
-    ParameterEnum,
-    Range,
+from envinorma.parametrization.models.condition import AndCondition, Equal, Greater, Littler, OrCondition, Range
+from envinorma.parametrization.models.parameter import ParameterEnum
+from envinorma.parametrization.natural_language_warnings import (
     _alineas_prefix,
     _generate_prefix,
     _is_range_of_size_at_least_2,
